feat:add new code

+++ src/modules/subcription/services/adminplan.service.ts
@@ -3,6 +3,7 @@ import { Stripe } from "../../../shared/facade";
 import { StripeCreatePlanInput, StripeDeletePlanInput } from "../../../shared/dto";
 import { PlanRepository } from "../../../shared/repositories";
 import { CreatePlanOutput } from "../dto";
+import { BadRequestError, InternalServerError } from "../../../shared/errors";
 
 @injectable()
 export default class AdminPlanService {
@@ -14,7 +15,17 @@ export default class AdminPlanService {
   async createPlan(args: StripeCreatePlanInput): Promise<CreatePlanOutput> {
     const { name, amount } = args;
 
+    const checkName = await this.planRepository.fetchOneByName(name);
+
+    if (checkName) {
+      throw new BadRequestError("name already exists");
+    }
+
     const stripeplan = await this.stripe.createPlan(args);
+
+    if (!stripeplan) {
+      throw new InternalServerError("failed to create plan on payment provider");
+    }
 
     const plan = await this.planRepository.create({
       name,
@@ -22,6 +33,10 @@ export default class AdminPlanService {
       stripe_plan_id: stripeplan.plan_id,
       stripe_price_id: stripeplan.price_id,
     });
+
+    if (!plan) {
+      throw new InternalServerError("failed to add plan to records");
+    }
 
     return {
       plan_id: plan.stripe_plan_id,
@@ -36,11 +51,23 @@ export default class AdminPlanService {
 
     const plan = await this.planRepository.fetchOneById(plan_id);
 
-    await this.stripe.deletePlan({
-      plan_id: plan.stripe_plan_id,
+    if (!plan) {
+      throw new BadRequestError("plan not found");
+    }
+
+    const stripeDelete = await this.stripe.deletePlan({
+      plan_id: plan!.stripe_plan_id,
     });
 
+    if (!stripeDelete) {
+      throw new InternalServerError("failed to delete plan on payment provider");
+    }
+
     const response = await this.planRepository.delete(plan_id);
+
+    if (!response) {
+      throw new InternalServerError("failed to delete plan from records");
+    }
     return {
       isDeleted: response ? true : false,
     };
